requirements.txt
Django=5.0.6
djangorestframework=3.15.1
drf-yasg==1.21.7
ujson==5.10.0
